Add CustomColor to ClientTheme Improve ClickGui scrolled

+++ src/main/java/keystrokesmod/utility/render/RenderUtils.java
@@ -504,10 +504,10 @@ public class RenderUtils {
         Tessellator tessellator = Tessellator.getInstance();
         WorldRenderer worldrenderer = tessellator.getWorldRenderer();
         worldrenderer.begin(7, DefaultVertexFormats.POSITION_COLOR);
-        worldrenderer.pos((double) right, (double) top, 0.0D).color(f1, f2, f3, f).endVertex();
-        worldrenderer.pos((double) left, (double) top, 0.0D).color(f1, f2, f3, f).endVertex();
-        worldrenderer.pos((double) left, (double) bottom, 0.0D).color(f5, f6, f7, f4).endVertex();
-        worldrenderer.pos((double) right, (double) bottom, 0.0D).color(f5, f6, f7, f4).endVertex();
+        worldrenderer.pos(right, top, 0.0D).color(f1, f2, f3, f).endVertex();
+        worldrenderer.pos(left, top, 0.0D).color(f1, f2, f3, f).endVertex();
+        worldrenderer.pos(left, bottom, 0.0D).color(f5, f6, f7, f4).endVertex();
+        worldrenderer.pos(right, bottom, 0.0D).color(f5, f6, f7, f4).endVertex();
         tessellator.draw();
         GlStateManager.shadeModel(7424);
         GlStateManager.disableBlend();
